+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+IN PROGRESS 0.2.0b0
+===================
+- Try to have always at least one TCP connection per host.
+- Quadratic backoff until reach the maximum of 60 seconds when an attempt for openning a connection
+  fails.
+
 0.1.1b0
 =======
 - Disabled support for `exptime` and `flags` for the `append` and `prepend` commands. For both commands